records: merge authors receivers
