+++ inspirehep/modules/authors/__init__.py
@@ -28,6 +28,4 @@ from __future__ import absolute_import, print_function
 
 from .ext import INSPIREAuthors
 
-from .receivers import *
-
 __all__ = ('INSPIREAuthors', )
